+++ src/main/java/com/africanb/africanb/dao/entity/security/Functionality.java
@@ -16,7 +16,7 @@ public class Functionality implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String code;
 
@@ -28,7 +28,7 @@ public class Functionality implements Serializable {
     private Date updatedAt;
     private Date deletedAt;
 
-    private Integer createdBy;
-    private Integer updatedBy;
-    private Integer deletedBy;
+    private Long createdBy;
+    private Long updatedBy;
+    private Long deletedBy;
 }
