+++ src/main/java/com/africanb/africanb/dao/entity/security/Role.java
@@ -14,7 +14,7 @@ public class Role implements Serializable {
 
     @Id @GeneratedValue(strategy = GenerationType.AUTO)
 
-    private Integer id;
+    private Long id;
     private String code;
     private String libelle;
 
@@ -24,7 +24,7 @@ public class Role implements Serializable {
     private Date updatedAt;
     private Date deletedAt;
 
-    private Integer createdBy;
-    private Integer updatedBy;
-    private Integer deletedBy;
+    private Long createdBy;
+    private Long updatedBy;
+    private Long deletedBy;
 }
