+++ src/main/java/com/africanb/africanb/dao/repository/security/FunctionalityRepository.java
@@ -15,7 +15,7 @@ import java.util.*;
 public interface FunctionalityRepository extends JpaRepository<Functionality, Integer> {
 
     @Query("select e from Functionality e where e.id= :id and e.isDeleted= :isDeleted")
-    Functionality findOne(@Param("id") Integer id, @Param("isDeleted") Boolean isDeleted);
+    Functionality findOne(@Param("id") Long id, @Param("isDeleted") Boolean isDeleted);
 
     @Query("select e from Functionality e where e.code= :code and e.isDeleted= :isDeleted")
     Functionality findByCode(@Param("code") String code, @Param("isDeleted") Boolean isDeleted);
