+++ src/main/java/com/africanb/africanb/dao/repository/security/RoleFunctionalityRepository.java
@@ -16,7 +16,7 @@ import java.util.*;
 public interface RoleFunctionalityRepository extends JpaRepository<RoleFunctionality, Integer> {
 
     @Query("select e from RoleFunctionality e where e.id= :id and e.isDeleted= :isDeleted")
-    RoleFunctionality findOne(@Param("id") Integer id, @Param("isDeleted") Boolean isDeleted);
+    RoleFunctionality findOne(@Param("id") Long id, @Param("isDeleted") Boolean isDeleted);
 
     @Query("select e from RoleFunctionality e where e.role.id= :roleId and e.functionality.id= :functionalityId and e.isDeleted= :isDeleted")
     RoleFunctionality findByRoleAndFunctionalityId(@Param("roleId") Integer roleId,@Param("functionalityId") Integer functionalityId, @Param("isDeleted") Boolean isDeleted);
@@ -25,10 +25,10 @@ public interface RoleFunctionalityRepository extends JpaRepository<RoleFunctiona
     RoleFunctionality findByRoleAndFunctionalityCode(@Param("roleCode") String roleCode,@Param("functionalityCode") String functionalityCode, @Param("isDeleted") Boolean isDeleted);
 
     @Query("select e from RoleFunctionality e where e.role.id= :roleId and e.isDeleted= :isDeleted")
-    List<RoleFunctionality> findByRoleId(@Param("roleId") Integer roleId, @Param("isDeleted") Boolean isDeleted);
+    List<RoleFunctionality> findByRoleId(@Param("roleId") Long roleId, @Param("isDeleted") Boolean isDeleted);
 
     @Query("select e from RoleFunctionality e where e.functionality.id= :functionalityId and e.isDeleted= :isDeleted")
-    List<RoleFunctionality> findByFunctionalityId(@Param("functionalityId") Integer functionalityId, @Param("isDeleted") Boolean isDeleted);
+    List<RoleFunctionality> findByFunctionalityId(@Param("functionalityId") Long functionalityId, @Param("isDeleted") Boolean isDeleted);
 
     @Query("select e from RoleFunctionality e where e.role.code= :roleCode and e.isDeleted= :isDeleted")
     List<RoleFunctionality> findByRoleCode(@Param("roleCode") String roleCode, @Param("isDeleted") Boolean isDeleted);
@@ -58,7 +58,7 @@ public interface RoleFunctionalityRepository extends JpaRepository<RoleFunctiona
     List<RoleFunctionality> findByDeletedBy(@Param("deletedBy")Long deletedBy, @Param("isDeleted")Boolean isDeleted);
 
     @Query("select e.functionality from RoleFunctionality e where e.role.id= :roleId and e.isDeleted= :isDeleted")
-    List<Functionality> findFunctionalityByRoleId(@Param("roleId") Integer roleId, @Param("isDeleted") Boolean isDeleted);
+    List<Functionality> findFunctionalityByRoleId(@Param("roleId") Long roleId, @Param("isDeleted") Boolean isDeleted);
 
     @Query("select e.functionality from RoleFunctionality e where e.role.code= :roleCode and e.isDeleted= :isDeleted")
     List<Functionality> findFunctionalityByRoleCode(@Param("roleCode") String roleCode, @Param("isDeleted") Boolean isDeleted);
